update_endlos

+++ Assets/Scripts/Weapon.cs
@@ -69,7 +69,7 @@ public class Weapon : MonoBehaviour
     /// </summary>
     /// <param name="additionalDmg">erhöhten den schaden der bullet direkt über diesen Wert</param>
     /// <param name="dmgModifier">nach hinzufügen des additionalDmg modifiers wir der dmg mit diesem Wert multipliziert</param>
-    public void shoot(int additionalDmg, float dmgModifier) {
+    public void shoot(float additionalDmg, float dmgModifier) {
         if (canShoot == true) {
             canShoot = false;
